Switch from Java UI code to FXML
