+++ pom.xml
@@ -98,7 +98,7 @@
             <!-- Default configuration for running with: mvn clean javafx:run -->
             <id>default-cli</id>
             <configuration>
-              <mainClass>com.turtlesamigo.lungabnormdetector/com.turtlesamigo.lungabnormdetector.HelloApplication</mainClass>
+              <mainClass>com.turtlesamigo.lungabnormdetector/com.turtlesamigo.lungabnormdetector.Main</mainClass>
               <launcher>app</launcher>
               <jlinkZipName>app</jlinkZipName>
               <jlinkImageName>app</jlinkImageName>
