+++ src/main/resources/com/turtlesamigo/lungabnormdetector/main-view.fxml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.Group?>
+<?import javafx.scene.Scene?>
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.VBox?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.image.ImageView?>
+<Group xmlns="http://javafx.com/javafx"
+       xmlns:fx="http://javafx.com/fxml"
+       fx:controller="com.turtlesamigo.lungabnormdetector.MainController">
+
+  <VBox>
+    <MenuBar>
+      <Menu text="File">
+        <MenuItem text="Open" onAction="#onOpenImage"/>
+        <SeparatorMenuItem/>
+        <MenuItem text="Exit" onAction="#exitApplication"/>
+      </Menu>
+    </MenuBar>
+
+    <ImageView fx:id="imageView"/>
+    <Label fx:id="fileNameLabel" text="File Name: "/>
+    <Label fx:id="fileSizeLabel" text="File Size: "/>
+  </VBox>
+</Group>
